+++ tests/tests_classifier/test_logistic_regression.py
@@ -14,21 +14,44 @@ X[:,1] = (X[:,1] - X[:,1].mean()) / X[:,1].std()
 
 
 def test_logistic_regression_gd():
-
-    t = np.array([ 0.52, 1.20, 4.40])
-    lr = LogisticRegression(epochs=100, eta=0.01, learning='gd')
+    t = np.array([ 0.55,  1.29,  4.39])
+    lr = LogisticRegression(epochs=100, eta=0.01, learning='gd', random_seed=0)
 
     lr.fit(X, y) # 0, 1 class
     np.testing.assert_almost_equal(lr.w_, t, 2)
     assert((y == lr.predict(X)).all())
-
 
 
 def test_logistic_regression_sgd():
-
-    t = np.array([ 0.51, 1.18, 4.38])
-    lr = LogisticRegression(epochs=100, eta=0.01, learning='sgd')
+    t = np.array([0.55,  1.26,  4.39])
+    lr = LogisticRegression(epochs=100, eta=0.01, learning='sgd', random_seed=0)
 
     lr.fit(X, y) # 0, 1 class
     np.testing.assert_almost_equal(lr.w_, t, 2)
     assert((y == lr.predict(X)).all())
+
+
+def test_l2_regularization_sgd():
+    lr = LogisticRegression(eta=0.01, epochs=20,
+                    learning='sgd', regularization='l2',
+                    lambda_=1.0, random_seed=0)
+    lr.fit(X, y)
+    y_pred = lr.predict(X)
+    expect_weights = np.array([0.100,  0.232,  0.346])
+
+    np.testing.assert_almost_equal(lr.w_, expect_weights, 3)
+    acc = sum(y_pred == y) / len(y)
+    assert(acc == 1.0)
+
+
+def test_l2_regularization_gd():
+    lr = LogisticRegression(eta=0.01, epochs=20,
+                    learning='gd', regularization='l2',
+                    lambda_=1.0, random_seed=0)
+    lr.fit(X, y)
+    y_pred = lr.predict(X)
+    expect_weights = np.array([ 0.252,  1.186,  2.296])
+
+    np.testing.assert_almost_equal(lr.w_, expect_weights, 3)
+    acc = sum(y_pred == y) / len(y)
+    assert(acc == 1.0)